Moves LaunchCommand class from Main to AbstractMain, fixing problems in derived classes (and removing circular dependency)
